Modify the logic to fit js-binding requirements.

+++ DragonBones/src/dragonBones/parsers/JSONDataParser.h
@@ -88,7 +88,7 @@ protected:
     {
         if (rawData.Size() > index)
         {
-            return rawData[index].GetInt();
+            return rawData[(int) index].GetInt();
         }
 
         return defaultValue;
@@ -98,7 +98,7 @@ protected:
     {
         if (rawData.Size() > index)
         {
-            return rawData[index].GetFloat();
+            return rawData[(int) index].GetFloat();
         }
 
         return defaultValue;
@@ -108,7 +108,7 @@ protected:
     {
         if (rawData.Size() > index)
         {
-            return rawData[index].GetString();
+            return rawData[(int) index].GetString();
         }
 
         return defaultValue;
